pyproject.toml
[project]
name = "gen-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.9.0",
    "isort>=6.1.0",
    "langchain>=0.3.27",
    "langchain-core>=0.3.79",
    "langchain-ollama>=0.3.10",
    "langchain-openai>=0.3.35",
    "langchain-tavily>=0.2.12",
    "langgraph>=1.0.0",
    "python-dotenv>=1.1.1",
]
